ingestion-service/requirements.txt
# Ingestion Service Dependencies
# Install these one by one and pin versions after installation

# Core framework
fastapi
uvicorn[standard]

# gRPC
grpcio
grpcio-tools

# Kafka
kafka-python

# Data validation
pydantic
pydantic-settings

# Authentication
python-jose[cryptography]
passlib[bcrypt]

# HTTP clients
aiohttp
httpx
websockets

# Crypto exchange library
ccxt

# Monitoring
prometheus-client

# Database
psycopg2-binary
sqlalchemy

# Utilities
python-dateutil

# Development tools
flake8
black
isort
pre-commit

# Testing
pytest
pytest-asyncio

